requirements.txt
aiohttp
beautifulsoup4
boto3
fastapi
feedparser
hdx-python-api
langchain_community
langchain_core
langchain_openai
langgraph
llama_index
pdfplumber
pandas
pydantic
python-dotenv
requests
seleniumbase
tenacity
uvicorn[standard]
websockets
yfinance

# Optional LlamaIndex integration (only needed if you use OpenAILike in rag_pipeline.py)
llama-index-llms-openai-like